+++ src/app/(pages)/_components/modal/index.tsx
@@ -1,1 +1,2 @@
-export * from './global-modal';
+export * from './auth';
+export * from './movie-select';
