+++ src/app/(pages)/_components/modal/movie-select/index.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import { HotMovieList } from '@/app/(pages)/_components/modal/movie-select/hot-movie-list';
-import { SearchContent } from '@/app/(pages)/_components/modal/movie-select/search-content';
 import { Dialog, DialogContent } from '@/components';
 import { close, selectModal } from '@/redux/features/modal-slice';
 import { useAppDispatch, useAppSelector } from '@/redux/hooks';
@@ -9,7 +7,10 @@ import { css, cx } from '@/styled-system/css';
 import { Flex } from '@/styled-system/jsx';
 import { flex } from '@/styled-system/patterns';
 
-export default function MovieSelectModal() {
+import { HotMovieList } from './hot-movie-list';
+import { SearchContent } from './search-content';
+
+export function MovieSelectModal() {
   const dispatch = useAppDispatch();
   const { isOpen, type } = useAppSelector(selectModal);
