+++ src/app/(pages)/_components/modal/movie-select/search-content-result.tsx
@@ -1,9 +1,6 @@
 'use client';
 
 import { getSearchData } from '@/api/search/get-search-data';
-import { MovieItem } from '@/app/(pages)/_components/modal/movie-select/movie-item';
-import { MovieListSkeleton } from '@/app/(pages)/_components/modal/movie-select/movie-list-skeleton';
-import { SearchContentNoResult } from '@/app/(pages)/_components/modal/movie-select/search-content-no-result';
 import { useObserverEffect } from '@/hooks';
 import { close } from '@/redux/features/modal-slice';
 import { useAppDispatch } from '@/redux/hooks';
@@ -11,6 +8,10 @@ import { css, cx } from '@/styled-system/css';
 import { flex, grid } from '@/styled-system/patterns';
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { Fragment, useRef } from 'react';
+
+import { MovieItem } from './movie-item';
+import { MovieListSkeleton } from './movie-list-skeleton';
+import { SearchContentNoResult } from './search-content-no-result';
 
 interface SearchContentResultProps {
   searchQuery: string;
