+++ src/app/(pages)/layout.tsx
@@ -1,20 +1,17 @@
-import { Footer, GlobalModal, Sidebar } from '@/app/(pages)/_components';
+import NextTopLoader from 'nextjs-toploader';
+
+import { AuthModal, Footer, MovieSelectModal, Sidebar } from '@/app/(pages)/_components';
 import { css } from '@/styled-system/css';
 import { Container } from '@/styled-system/jsx';
 import { token } from '@/styled-system/tokens';
 import { Metadata } from 'next';
-import NextTopLoader from 'nextjs-toploader';
 
 export const metadata: Metadata = {
   title: '씨네필',
   description: '영화를 좋아하는 사람들을 위한 커뮤니티',
 };
 
-interface PagesLayoutProps {
-  children: React.ReactNode;
-}
-
-export default function PagesLayout({ children }: PagesLayoutProps) {
+export default function PagesLayout({ children }: { children: React.ReactNode }) {
   return (
     <>
       <NextTopLoader color={token('colors.gray.50')} showSpinner={false} height={1} shadow={false} />
@@ -23,7 +20,8 @@ export default function PagesLayout({ children }: PagesLayoutProps) {
         <main className={css({ flex: 1, h: 'full', ml: { base: 0, md: 64 }, mb: 16 })}>{children}</main>
         <Footer />
       </Container>
-      <GlobalModal />
+      <MovieSelectModal />
+      <AuthModal />
     </>
   );
 }
